Added a cast to `int16' from the result of `Dftype' to remove a warning.
git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@15784 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/bench/C/parsing/shared/yacc.c
@@ -424,13 +424,13 @@ char is_feature_as(char *obj)
 {
 	/* Is `obj' an instance of FEATURE_AS ? */
 
-	return (Dftype(obj) == feature_as_dtype) ? '\01' : '\0';
+	return ((int16) Dftype(obj) == feature_as_dtype) ? '\01' : '\0';
 }
 
 char is_invariant_as(char *obj)
 {
 	/* Is `obj' an instance of INVARIANT_AS ? */
 
-    return (Dftype(obj) == invariant_as_dtype) ? '\01' : '\0';
+    return ((int16) Dftype(obj) == invariant_as_dtype) ? '\01' : '\0';
 }
